+++ index.js
@@ -1,30 +1,15 @@
-function hexToRgb(hex) {
-  var result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
-  return result ? {
-    r: parseInt(result[1], 16),
-    g: parseInt(result[2], 16),
-    b: parseInt(result[3], 16)
-  } : null;
-}
-
-function componentToHex(c) {
-    var hex = c.toString(16);
-    return hex.length == 1 ? "0" + hex : hex;
-  }
-  
-  function rgbToHex(r, g, b) {
-    return "#" + componentToHex(r) + componentToHex(g) + componentToHex(b);
-  }
-  
-  ;
+const redGreenBlue = require('./rgb-to-hex')
+const hexCode = require('./hex-to-rgb')
 
 if (process.argv[2] === "hex") {
-    console.log(hexToRgb(process.argv[3]))  
-} else if (process.argv[2] === "rgb"){
+    console.log(hexCode.hexToRgb(process.argv[3]))
+
+} else if (process.argv[2] === "rgb") {
     let r = parseInt(process.argv[3])
     let g = parseInt(process.argv[4])
     let b = parseInt(process.argv[5])
-    
-    let rgb = rgbToHex(r,g,b)
+
+    let rgb = redGreenBlue.rgbToHex(r, g, b)
+
     console.log(rgb)
 }
